Workaround failing test in JS-legacy, update JS API dumps

+++ libraries/stdlib/api/js-v1/kotlin.kt
@@ -450,7 +450,7 @@ public inline fun <R, T> kotlin.Result<T>.mapCatching(transform: (value: T) -> R
 
 @kotlin.SinceKotlin(version = "1.5")
 @kotlin.internal.InlineOnly
-public inline fun kotlin.Byte.mod(other: kotlin.Byte): kotlin.Int
+public inline fun kotlin.Byte.mod(other: kotlin.Byte): kotlin.Byte
 
 @kotlin.SinceKotlin(version = "1.5")
 @kotlin.internal.InlineOnly
@@ -462,11 +462,11 @@ public inline fun kotlin.Byte.mod(other: kotlin.Long): kotlin.Long
 
 @kotlin.SinceKotlin(version = "1.5")
 @kotlin.internal.InlineOnly
-public inline fun kotlin.Byte.mod(other: kotlin.Short): kotlin.Int
+public inline fun kotlin.Byte.mod(other: kotlin.Short): kotlin.Short
 
 @kotlin.SinceKotlin(version = "1.5")
 @kotlin.internal.InlineOnly
-public inline fun kotlin.Int.mod(other: kotlin.Byte): kotlin.Int
+public inline fun kotlin.Int.mod(other: kotlin.Byte): kotlin.Byte
 
 @kotlin.SinceKotlin(version = "1.5")
 @kotlin.internal.InlineOnly
@@ -478,15 +478,15 @@ public inline fun kotlin.Int.mod(other: kotlin.Long): kotlin.Long
 
 @kotlin.SinceKotlin(version = "1.5")
 @kotlin.internal.InlineOnly
-public inline fun kotlin.Int.mod(other: kotlin.Short): kotlin.Int
+public inline fun kotlin.Int.mod(other: kotlin.Short): kotlin.Short
 
 @kotlin.SinceKotlin(version = "1.5")
 @kotlin.internal.InlineOnly
-public inline fun kotlin.Long.mod(other: kotlin.Byte): kotlin.Long
+public inline fun kotlin.Long.mod(other: kotlin.Byte): kotlin.Byte
 
 @kotlin.SinceKotlin(version = "1.5")
 @kotlin.internal.InlineOnly
-public inline fun kotlin.Long.mod(other: kotlin.Int): kotlin.Long
+public inline fun kotlin.Long.mod(other: kotlin.Int): kotlin.Int
 
 @kotlin.SinceKotlin(version = "1.5")
 @kotlin.internal.InlineOnly
@@ -494,11 +494,11 @@ public inline fun kotlin.Long.mod(other: kotlin.Long): kotlin.Long
 
 @kotlin.SinceKotlin(version = "1.5")
 @kotlin.internal.InlineOnly
-public inline fun kotlin.Long.mod(other: kotlin.Short): kotlin.Long
+public inline fun kotlin.Long.mod(other: kotlin.Short): kotlin.Short
 
 @kotlin.SinceKotlin(version = "1.5")
 @kotlin.internal.InlineOnly
-public inline fun kotlin.Short.mod(other: kotlin.Byte): kotlin.Int
+public inline fun kotlin.Short.mod(other: kotlin.Byte): kotlin.Byte
 
 @kotlin.SinceKotlin(version = "1.5")
 @kotlin.internal.InlineOnly
@@ -510,7 +510,7 @@ public inline fun kotlin.Short.mod(other: kotlin.Long): kotlin.Long
 
 @kotlin.SinceKotlin(version = "1.5")
 @kotlin.internal.InlineOnly
-public inline fun kotlin.Short.mod(other: kotlin.Short): kotlin.Int
+public inline fun kotlin.Short.mod(other: kotlin.Short): kotlin.Short
 
 @kotlin.internal.InlineOnly
 @kotlin.SinceKotlin(version = "1.3")
@@ -2438,7 +2438,7 @@ public final inline class UByte : kotlin.Comparable<kotlin.UByte> {
     public final inline operator fun minus(other: kotlin.UShort): kotlin.UInt
 
     @kotlin.internal.InlineOnly
-    public final inline fun mod(other: kotlin.UByte): kotlin.UInt
+    public final inline fun mod(other: kotlin.UByte): kotlin.UByte
 
     @kotlin.internal.InlineOnly
     public final inline fun mod(other: kotlin.UInt): kotlin.UInt
@@ -2447,7 +2447,7 @@ public final inline class UByte : kotlin.Comparable<kotlin.UByte> {
     public final inline fun mod(other: kotlin.ULong): kotlin.ULong
 
     @kotlin.internal.InlineOnly
-    public final inline fun mod(other: kotlin.UShort): kotlin.UInt
+    public final inline fun mod(other: kotlin.UShort): kotlin.UShort
 
     @kotlin.internal.InlineOnly
     public final inline infix fun or(other: kotlin.UByte): kotlin.UByte
@@ -2633,7 +2633,7 @@ public final inline class UInt : kotlin.Comparable<kotlin.UInt> {
     public final inline operator fun minus(other: kotlin.UShort): kotlin.UInt
 
     @kotlin.internal.InlineOnly
-    public final inline fun mod(other: kotlin.UByte): kotlin.UInt
+    public final inline fun mod(other: kotlin.UByte): kotlin.UByte
 
     @kotlin.internal.InlineOnly
     public final inline fun mod(other: kotlin.UInt): kotlin.UInt
@@ -2642,7 +2642,7 @@ public final inline class UInt : kotlin.Comparable<kotlin.UInt> {
     public final inline fun mod(other: kotlin.ULong): kotlin.ULong
 
     @kotlin.internal.InlineOnly
-    public final inline fun mod(other: kotlin.UShort): kotlin.UInt
+    public final inline fun mod(other: kotlin.UShort): kotlin.UShort
 
     @kotlin.internal.InlineOnly
     public final inline infix fun or(other: kotlin.UInt): kotlin.UInt
@@ -2834,16 +2834,16 @@ public final inline class ULong : kotlin.Comparable<kotlin.ULong> {
     public final inline operator fun minus(other: kotlin.UShort): kotlin.ULong
 
     @kotlin.internal.InlineOnly
-    public final inline fun mod(other: kotlin.UByte): kotlin.ULong
+    public final inline fun mod(other: kotlin.UByte): kotlin.UByte
 
     @kotlin.internal.InlineOnly
-    public final inline fun mod(other: kotlin.UInt): kotlin.ULong
+    public final inline fun mod(other: kotlin.UInt): kotlin.UInt
 
     @kotlin.internal.InlineOnly
     public final inline fun mod(other: kotlin.ULong): kotlin.ULong
 
     @kotlin.internal.InlineOnly
-    public final inline fun mod(other: kotlin.UShort): kotlin.ULong
+    public final inline fun mod(other: kotlin.UShort): kotlin.UShort
 
     @kotlin.internal.InlineOnly
     public final inline infix fun or(other: kotlin.ULong): kotlin.ULong
@@ -3035,7 +3035,7 @@ public final inline class UShort : kotlin.Comparable<kotlin.UShort> {
     public final inline operator fun minus(other: kotlin.UShort): kotlin.UInt
 
     @kotlin.internal.InlineOnly
-    public final inline fun mod(other: kotlin.UByte): kotlin.UInt
+    public final inline fun mod(other: kotlin.UByte): kotlin.UByte
 
     @kotlin.internal.InlineOnly
     public final inline fun mod(other: kotlin.UInt): kotlin.UInt
@@ -3044,7 +3044,7 @@ public final inline class UShort : kotlin.Comparable<kotlin.UShort> {
     public final inline fun mod(other: kotlin.ULong): kotlin.ULong
 
     @kotlin.internal.InlineOnly
-    public final inline fun mod(other: kotlin.UShort): kotlin.UInt
+    public final inline fun mod(other: kotlin.UShort): kotlin.UShort
 
     @kotlin.internal.InlineOnly
     public final inline infix fun or(other: kotlin.UShort): kotlin.UShort
